bamnado::bam_utils

Function convert_lapper_chrom_names_to_ids

pub fn convert_lapper_chrom_names_to_ids(
    lapper: HashMap<String, Lapper<usize, u32>>,
    bam_stats: &BamStats,
) -> Result<HashMap<usize, Lapper<usize, u32>>>

Converts the chromosome names in a Lapper to chromosome IDs based on a BAM file’s header.

This is useful for when we want to use the lapper with the BAM file.